+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.20.8 (April 26, 2024)
+
+### Fixed
+
+- Fix check for `Elemenent.prototype.checkVisibility` function
+- Fix filter animations in Safari
+
 ## 3.20.7 (April 24, 2024)
 
 ### Fixed